// types/slides.ts
export type ChartType = 'bar' | 'horizontalBar' | 'line' | 'pie' | 'ring' | 'area' | 'scatter'
export interface ChartOptions {
  lineSmooth?: boolean
  stack?: boolean
}
export interface ChartData {
  labels: string[]
  legends: string[]
  series: number[][]
}